Refaction Keep Token Transfer actions
