+++ solidity/dashboard/src/sagas/keep-balance.js
@@ -1,7 +1,15 @@
-import { takeLatest, call, put, retry } from "redux-saga/effects"
+import {
+  takeLatest,
+  call,
+  put,
+  retry,
+  actionChannel,
+  take,
+} from "redux-saga/effects"
 import { logError } from "./utils"
 import { getWeb3Context } from "./utils"
 import keepToken from "../services/keepToken"
+import { keepBalanceActions } from "../actions"
 
 export function* watchKeepTokenBalanceRequest() {
   yield takeLatest("keep-token/balance_request", fetchKeepTokenBalance)
@@ -27,5 +35,47 @@ function* fetchKeepTokenBalance() {
     })
   } catch (error) {
     yield* logError("keep-token/balance_request_failure", error)
+  }
+}
+
+export function* subscribeToKeepTokenTransferFromEvent() {
+  const requestChan = yield actionChannel(
+    keepBalanceActions.KEEP_TOKEN_TRANSFER_FROM_EVENT_EMITTED
+  )
+
+  // Observe and dispatch an action that updates keep token balance.
+  while (true) {
+    const {
+      payload: { event },
+    } = yield take(requestChan)
+    const {
+      returnValues: { value },
+    } = event
+
+    yield put({
+      type: "keep-token/transferred_from",
+      payload: { value },
+    })
+  }
+}
+
+export function* subscribeToKeepTokenTransferToEvent() {
+  const requestChan = yield actionChannel(
+    keepBalanceActions.KEEP_TOKEN_TRANSFER_TO_EVENT_EMITTED
+  )
+
+  // Observe and dispatch an action that updates keep token balance.
+  while (true) {
+    const {
+      payload: { event },
+    } = yield take(requestChan)
+    const {
+      returnValues: { value },
+    } = event
+
+    yield put({
+      type: "keep-token/transferred_to",
+      payload: { value },
+    })
   }
 }
